backend/requirements.txt
Django==5.0.4  # Latest stable Django (or 4.2.11 for LTS)
djangorestframework==3.15.1
mysqlclient==2.2.1
Pillow==10.2.0  # Must use Pillow 10+ for Python 3.13